Extract sms, ivr and mobileweb simulators to distinct files
